# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "DraftOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "material",
              "questions",
              "flashcards"
            ],
            "title": "Kind",
            "type": "string"
          },
          "module_id": {
            "format": "uuid",
            "title": "Module Id",
            "type": "string"
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          },
          "request": {
            "additionalProperties": true,
            "title": "Request",
            "type": "object"
          },
          "status": {
            "enum": [
              "generating",
              "ready",
              "failed",
              "saved",
              "discarded"
            ],
            "title": "Status",
            "type": "string"
          },
          "topic_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "module_id",
          "topic_id",
          "kind",
          "request",
          "status",
          "payload",
          "error_code",
          "created_at",
          "updated_at"
        ],
        "title": "DraftOut",
        "type": "object"
      },
      "DraftRegenerate": {
        "additionalProperties": false,
        "properties": {
          "instructions": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          }
        },
        "title": "DraftRegenerate",
        "type": "object"
      },
      "DraftSave": {
        "additionalProperties": false,
        "properties": {
          "content_md": {
            "anyOf": [
              {
                "maxLength": 200000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Md"
          },
          "selected": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "DraftSave",
        "type": "object"
      },
...
      "GenerateRequest": {
        "additionalProperties": false,
        "description": "Ask Claude for a material, questions or flashcards. The result is a\ndraft to preview before anything is saved.",
        "properties": {
          "count": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count"
          },
          "difficulty": {
            "anyOf": [
              {
                "enum": [
                  "easy",
                  "medium",
                  "hard",
                  "exam"
                ],
                "type": "string"
              },
              {
                "const": "mixed",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty"
          },
          "document_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 10,
            "title": "Document Ids",
            "type": "array"
          },
          "improve_material_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Improve Material Id"
          },
          "instructions": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "kind": {
            "enum": [
              "material",
              "questions",
              "flashcards"
            ],
            "title": "Kind",
            "type": "string"
          },
          "material_kind": {
            "anyOf": [
              {
                "enum": [
                  "guide",
                  "summary",
                  "formula_sheet",
                  "worked_examples",
                  "definitions",
                  "explanation",
                  "concept_map",
                  "notes"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Kind"
          },
          "module_id": {
            "format": "uuid",
            "title": "Module Id",
            "type": "string"
          },
          "topic_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          },
          "types": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "multiple_choice",
                    "true_false",
                    "numerical",
                    "expression",
                    "short_answer",
                    "explanation",
                    "derivation"
                  ],
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Types"
          }
        },
        "required": [
          "module_id",
          "kind"
        ],
        "title": "GenerateRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SavedDraft": {
        "properties": {
          "draft": {
            "$ref": "#/components/schemas/DraftOut"
          },
          "material_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Id"
          },
          "saved_items": {
            "default": 0,
            "title": "Saved Items",
            "type": "integer"
          }
        },
        "required": [
          "draft"
        ],
        "title": "SavedDraft",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Revision OS",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/drafts": {
      "get": {
        "description": "Drafts not yet saved or discarded.",
        "operationId": "list_drafts_api_v1_drafts_get",
        "parameters": [
          {
            "in": "query",
            "name": "module_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Module Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DraftOut"
                  },
                  "title": "Response List Drafts Api V1 Drafts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Drafts",
        "tags": [
          "materials"
        ]
      },
      "post": {
        "description": "Ask Claude for a material, questions or flashcards. Poll the draft\nuntil it is ready, then save, regenerate or discard it.",
        "operationId": "generate_api_v1_drafts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate",
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/drafts/{draft_id}": {
      "get": {
        "operationId": "get_draft_api_v1_drafts__draft_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Draft",
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/drafts/{draft_id}/discard": {
      "post": {
        "operationId": "discard_draft_api_v1_drafts__draft_id__discard_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discard Draft",
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/drafts/{draft_id}/regenerate": {
      "post": {
        "operationId": "regenerate_draft_api_v1_drafts__draft_id__regenerate_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftRegenerate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Regenerate Draft",
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/drafts/{draft_id}/save": {
      "post": {
        "operationId": "save_draft_api_v1_drafts__draft_id__save_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftSave"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedDraft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Draft",
        "tags": [
          "materials"
        ]
      }
    },
...
    "/api/v1/materials": {
      "get": {
        "operationId": "list_materials_api_v1_materials_get",
        "parameters": [
          {
            "in": "query",
            "name": "module_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Module Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MaterialSummary"
                  },
                  "title": "Response List Materials Api V1 Materials Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Materials",
        "tags": [
          "materials"
        ]
      },
      "post": {
        "description": "One of your own materials (origin: user).",
        "operationId": "create_material_api_v1_materials_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Material",
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/materials/{material_id}": {
      "delete": {
        "operationId": "delete_material_api_v1_materials__material_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "material_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Material Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Material",
        "tags": [
          "materials"
        ]
      },
      "get": {
        "operationId": "get_material_api_v1_materials__material_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "material_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Material Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Material",
        "tags": [
          "materials"
        ]
      },
      "patch": {
        "operationId": "update_material_api_v1_materials__material_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "material_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Material Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Material",
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/materials/{material_id}/diff": {
      "get": {
        "operationId": "diff_versions_api_v1_materials__material_id__diff_get",
        "parameters": [
          {
            "in": "path",
            "name": "material_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Material Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from_version",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "From Version",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to_version",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "To Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiffOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Diff Versions",
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/materials/{material_id}/restore": {
      "post": {
        "operationId": "restore_material_api_v1_materials__material_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "material_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Material Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Material",
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/materials/{material_id}/versions": {
      "post": {
        "description": "Your edit, saved as a new version (the old ones are kept).",
        "operationId": "add_version_api_v1_materials__material_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "material_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Material Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Version",
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/materials/{material_id}/versions/{version_id}": {
      "delete": {
        "operationId": "delete_version_api_v1_materials__material_id__versions__version_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "material_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Material Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Version",
        "tags": [
          "materials"
        ]
      },
      "get": {
        "operationId": "get_version_api_v1_materials__material_id__versions__version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "material_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Material Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Version",
        "tags": [
          "materials"
        ]
      }
    },
    "/api/v1/materials/{material_id}/versions/{version_id}/restore": {
      "post": {
        "operationId": "restore_version_api_v1_materials__material_id__versions__version_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "material_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Material Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Version",
        "tags": [
          "materials"
        ]
      }
    },
...
  }
}